requirements.txt
# Fire Detection System - Requirements

Django==4.2.5
djangorestframework==3.14.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
ultralytics==8.0.196
torch==2.0.1
torchvision==0.15.2
python-decouple==3.8
celery==5.3.4
redis==5.0.1
requests==2.31.0
